Move new readers
